fix: scores ok with coeffs, was problem in db with scoremin and max as integer whereas score was a float

+++ back-server/database/cms-database.sql
@@ -194,8 +194,8 @@ CREATE TABLE Scores (
 CREATE TABLE JoinScoresQuestionnaires (
     questionnaire_id INTEGER NOT NULL,
     score_id INTEGER NOT NULL,
-    scoremax INTEGER NOT NULL,
-    scoremin INTEGER NOT NULL,
+    scoremax FLOAT NOT NULL,
+    scoremin FLOAT NOT NULL,
     PRIMARY KEY (questionnaire_id, score_id),
     FOREIGN KEY (questionnaire_id) REFERENCES Questionnaires(id) ON DELETE CASCADE,
     FOREIGN KEY (score_id) REFERENCES Scores(id) ON DELETE CASCADE
@@ -221,3 +221,5 @@ INSERT INTO Scores (lettre) VALUES
 ('A'),
 ('B'),
 ('C');
+
+
